Norminetting everything to make it easier to add things.
